examples/foc: move the parse args logic to a separate file
